refactor(app): extract project I/O into useProjectIO hook
Move loadProject, handleSaveProjectSettings, handleSaveAll,
handleReloadFromDisk, and handleRefreshProject out of App.tsx into a
dedicated src/hooks/useProjectIO.ts hook, following the established
src/hooks/ extraction pattern. App.tsx drops from 5314 to 4598 lines.

Part of the App.tsx maintainability refactor (bmf-vangard-renpy-ide-cfo).
