+++ .beads/interactions.jsonl
@@ -24,3 +24,4 @@
 {"id":"int-07cf3e7d","kind":"field_change","created_at":"2026-06-14T18:56:22.8848685Z","actor":"G.Hirpara","issue_id":"bmf-vangard-renpy-ide-73l","extra":{"field":"status","new_value":"closed","old_value":"in_progress","reason":"Closed"}}
 {"id":"int-d7428a6d","kind":"field_change","created_at":"2026-06-14T19:09:52.8557923Z","actor":"G.Hirpara","issue_id":"bmf-vangard-renpy-ide-zk1","extra":{"field":"status","new_value":"closed","old_value":"in_progress","reason":"Closed"}}
 {"id":"int-7fec418f","kind":"field_change","created_at":"2026-06-23T01:34:08.7056085Z","actor":"G.Hirpara","issue_id":"bmf-vangard-renpy-ide-hj5","extra":{"field":"status","new_value":"closed","old_value":"in_progress","reason":"Added 3 missing smoke flows: run+stop game (06), edit+save+reload (07), and diagnostics clickthrough navigation (extended 04). Fixed fake renpy.sh to actually sleep/block so run/stop is testable on Linux CI; compiled a matching renpy.exe stub for Windows local dev. Added a deliberate invalid jump to the fixture script.rpy so diagnostics has a real clickable issue. All 11 smoke tests pass locally."}}
+{"id":"int-d208ad6f","kind":"field_change","created_at":"2026-06-23T03:01:32.2470042Z","actor":"G.Hirpara","issue_id":"bmf-vangard-renpy-ide-gn2","extra":{"field":"status","new_value":"closed","old_value":"in_progress","reason":"Extracted loadProject/handleSaveProjectSettings/handleSaveAll/handleReloadFromDisk/handleRefreshProject into src/hooks/useProjectIO.ts. App.tsx 5314->4598 lines (-716). tsc/lint/686 tests/build all green, no regressions (verified via before/after diff of pre-existing tsc errors)."}}
